final fix

+++ modules/db/migrations/init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS user_subscriptions (
-    user_id INT UNIQUE,
+    user_id BIGINT UNIQUE,
     subscription_type INT,
     start_date DATE,
     end_date DATE
